pyproject.toml
[tool.poetry]
name = "price-checker"
version = "0.1.0"
description = ""
authors = ["ivanmikhailov <mikhailovj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
selenium = "^4.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
